LayoutTests/media/video-source-add-src-expected.txt

EVENT(loadstart)
EVENT(loadstart)

EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
RUN(video.src = 'content/silence.mpg')

EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)

EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
END OF TEST
Test that adding a 'src' attribute does nothing when a <source> element has already been chosen
